==> Lessons/06_DataWrangling.Rmd <==
---
title: "6: Data Wrangling"
author: "Environmental Data Analytics | Kateri Salk"
date: "Spring 2020"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1. Describe the usefulness of data wrangling and its place in the data pipeline
2. Wrangle datasets with dplyr functions
3. Apply data wrangling skills to a real-world example dataset

## Set up your session

Today we will work with a dataset from the [North Temperate Lakes Long-Term Ecological Research Station](https://lter.limnology.wisc.edu/about/overview). The NTL-LTER is located in the boreal zone in northern Wisconsin, USA. We will use the [chemical and physical limnology dataset](https://lter.limnology.wisc.edu/content/cascade-project-north-temperate-lakes-lter-core-data-physical-and-chemical-limnology-1984), running from 1984-2016. 

Opening discussion: why might we be interested in long-term observations of temperature, oxygen, and light in lakes?

> Add notes here: Ecosystem health, climate change effects, range of tolerance for organisms, management techniques.

```{r, message = FALSE}
getwd()
library(tidyverse)
install.packages(lubridate)
library(lubridate)
NTL.phys.data <- read.csv("./Data/Raw/NTL-LTER_Lake_ChemistryPhysics_Raw.csv")

colnames(NTL.phys.data)
head(NTL.phys.data)
# gives first 6 rows of each of the columns in that data frame
summary(NTL.phys.data)
str(NTL.phys.data)
dim(NTL.phys.data)

class(NTL.phys.data$sampledate)
# Format sampledate as date
NTL.phys.data$sampledate <- as.Date(NTL.phys.data$sampledate, format = "%m/%d/%y")
#format = current format the date is in. 
```

## Data Wrangling

Data wrangling extends data exploration: it allows you to process data in ways that are useful for you. An important part of data wrangling is creating *tidy datasets*, with the following rules: 

1. Each variable has its own column
2. Each observation has its own row
3. Each value has its own cell

What is the best way to wrangle data? There are multiple ways to arrive at a specific outcome in R, and we will illustrate some of those approaches. Your goal should be to write the simplest code that will get you to your desired outcome. However, there is sometimes a trade-off of the opportunity cost to learn a new formulation of code and the time it takes to write complex code that you already know. Remember that the best code is one that is easy to understand for yourself and your collaborators. Remember to comment your code, use informative names for variables and functions, and use reproducible methods to arrive at your output.

## Dplyr Wrangling Functions

`dplyr` is a package in R that includes functions for data manipulation (i.e., data wrangling or data munging). `dplyr` is included in the tidyverse package, so you should already have it installed on your machine. The functions act as verbs for data wrangling processes. For more information, run this line of code:

```{r, results = "hide"}
vignette("dplyr")

# Results = hide means that the entire vignette won't show up
```

### Filter

Filtering allows us to choose certain rows (observations) in our dataset.

Here are the relevant commands used in the `filter` function. Add some notes to designate what these commands mean. 
`==` #Equals. Has to be 2 on filter function
`!=` # Not equal to
`<` #Less than
`<=` #Less than or equal to
`>` # Greater than
`>=` #Greater than or equal to
`&` # And
`|` # Or

```{r}
class(NTL.phys.data$lakeid)
class(NTL.phys.data$depth)

# matrix filtering
NTL.phys.data.surface1 <- NTL.phys.data[NTL.phys.data$depth == 0,]
# If want to make a new data frame, write surface 1, then on other side select the depth of 0 to only see data with depth of 0
# In matrix notation, have to have comma, so can leave either side blank but have to include it

# dplyr filtering
NTL.phys.data.surface2 <- filter(NTL.phys.data, depth == 0)
# Notice this has the same dimensions as matrix above
NTL.phys.data.surface3 <- filter(NTL.phys.data, depth < 0.25)
# filter function, include: which data to grab from and how to do the filtering
# note: double equals
# tibbles can be subed out for data frame. have slight differences, but not important right now

# Did the methods arrive at the same result?
head(NTL.phys.data.surface1)
dim(NTL.phys.data.surface1)
head(NTL.phys.data.surface2)
dim(NTL.phys.data.surface2)
head(NTL.phys.data.surface3)
dim(NTL.phys.data.surface3)

# Choose multiple conditions to filter
summary(NTL.phys.data$lakename)
# Tells you the number of rows contain each lake name
NTL.phys.data.PeterPaul1 <- filter(NTL.phys.data, lakename == "Paul Lake" | lakename == "Peter Lake")
NTL.phys.data.PeterPaul2 <- filter(NTL.phys.data, lakename != "Central Long Lake" & 
                                     lakename != "Crampton Lake" & lakename != "East Long Lake" &
                                     lakename != "Hummingbird Lake" & lakename != "Tuesday Lake" &
                                     lakename != "Ward Lake" & lakename != "West Long Lake")

NTL.phys.data.PeterPaul3 <- filter(NTL.phys.data, lakename %in% c("Paul Lake", "Peter Lake"))
# I want the new dataset to include anything that is Paul and Peter Lake. Note "c" to concatinate

# Choose a range of conditions of a numeric or integer variable
summary(NTL.phys.data$daynum)
# 55 = 55th day of the year, day 307: near the end of the year. Not sampling when the lake is frozen (Very common where there's ice that samples are concentrated in the warmer seasons)
NTL.phys.data.JunethruOctober1 <- filter(NTL.phys.data, daynum > 151 & daynum < 305)
# If want to choose the dates from June to October, can filter the data set and want filter to filter numbers greater than 151 and less than 305
NTL.phys.data.JunethruOctober2 <- filter(NTL.phys.data, daynum > 151, daynum < 305)
NTL.phys.data.JunethruOctober3 <- filter(NTL.phys.data, daynum >= 152 & daynum <= 304)
NTL.phys.data.JunethruOctober4 <- filter(NTL.phys.data, daynum %in% c(152:304))
# All of these come out to the same dimensions

# Exercise: 
# filter NTL.phys.data for the year 1999
# what code do you need to use, based on the class of the variable?
class(NTL.phys.data$year4) #integer
NTL.phys.data.1999 <- filter(NTL.phys.data, year4 == 1999)
dim(NTL.phys.data.1999)

# Exercise: 
# filter NTL.phys.data for Tuesday Lake from 1990 through 1999.
NTL.phys.data.TuesdayLake.90thru99 <- filter(NTL.phys.data, lakename == "Tuesday Lake" & year4 >1989, year4 <2000)
dim(NTL.phys.data.TuesdayLake.90thru99)

# OR

NTL.phys.data.Tuesday <- filter(NTL.phys.data,
                                lakename == "Tuesday Lake" &
                                  year4 >1989 & year4 <2000) 

dim(NTL.phys.data.Tuesday)

```
Question: Why don't we filter using row numbers?

> Answer: More work and if dataset changed, the row numbers would change.

### Arrange

Arranging allows us to change the order of rows in our dataset. By default, the arrange function will arrange rows in ascending order.

```{r}
NTL.phys.data.depth.ascending <- arrange(NTL.phys.data, depth)
# Arrange physical dataset by ascending depth. Don't have to specify, but have to specify for descending
NTL.phys.data.depth.descending <- arrange(NTL.phys.data, desc(depth))
# Arrange physical dataset by descending depth.

# Exercise: 
# Arrange NTL.phys.data by temperature, in descending order. 

NTL.phys.data.Temp.Desc <- arrange(NTL.phys.data, desc(temperature_C))

# Which dates, lakes, and depths have the highest temperatures?
# Answer: June and July, East Long Lake and Hummingbird Lake, 1995-2002, at shallow depths

# Can temporarily change via environment by clicking arrow on column name to change ascending/descending. But this doesn't change it permanently
```
### Select

Selecting allows us to choose certain columns (variables) in our dataset.

```{r}
NTL.phys.data.temps <- select(NTL.phys.data, lakename, sampledate:temperature_C)

# Can just put in commas, don't have to add logical operators like in filter
# Name which column we want to keep or leave out of dataset. Colon = Sample Date THROUGH Temp. 
# Different from filter function because in filter, can get certain attributes, but in select function, we're only selecting for a certain column name. Also, filter is manipulating the rows (horizontal) and selecting is always columns (vertical)

#Can make it not include a certain column by adding a minus (-) before the name

NTL.phys.data.temps.not.included <- select(NTL.phys.data, -lakename, sampledate:temperature_C)


```
### Mutate

Mutating allows us to add new columns that are functions of existing columns. Operations include addition, subtraction, multiplication, division, log, and other functions.

```{r}

NTL.phys.data.temps <- mutate(NTL.phys.data.temps, temperature_F = (temperature_C*9/5) + 32)

# Want to add a column, call it "temp_F", and say what info u want in it. Where ever we have temp in C, it changes it to F and adds it to new column
# Can't choose location of where it adds it. Always adds it to far right.
# To rearrange it, use select function and call out certain columns in the order you want it.
```

## Lubridate

A package that makes coercing date much easier is `lubridate`. A guide to the package can be found at https://lubridate.tidyverse.org/. The cheat sheet within that web page is excellent too. This package can do many things (hint: look into this package if you are having unique date-type issues), but today we will be using two of its functions for our NTL dataset. 

```{r}
# SUPER useful if dealing with dates in R

# add a month column to the dataset
NTL.phys.data.PeterPaul1 <- mutate(NTL.phys.data.PeterPaul1, month = month(sampledate)) 
# Creates new column called month. Lubridate has a function named month so it's easy. Ask it to determine which month was the sample date.
# Automatically puts month column at the end

# reorder columns to put month with the rest of the date variables
NTL.phys.data.PeterPaul1 <- select(NTL.phys.data.PeterPaul1, lakeid:daynum, month, sampledate:comments)
# Select items from PeterPaul1 and then just put them in different order. Lake ID thru Day Number, then month, then sample date thru comments.

# find out the start and end dates of the dataset. Can find out intervals between certain dates. Can ask specifically for interval between first and last dates. When did we start sampling? WHen did we end? See Below:

interval(NTL.phys.data.PeterPaul1$sampledate[1], NTL.phys.data.PeterPaul1$sampledate[21613])
interval(first(NTL.phys.data.PeterPaul1$sampledate), last(NTL.phys.data.PeterPaul1$sampledate))
```


## Pipes

Sometimes we will want to perform multiple functions on a single dataset on our way to creating a processed dataset. We could do this in a series of subsequent functions or create a custom function. However, there is another method to do this that looks cleaner and is easier to read. This method is called a pipe. We designate a pipe with `%>%`. A good way to think about the function of a pipe is with the word "then." 

Let's say we want to take our raw dataset (NTL.phys.data), *then* filter the data for Peter and Paul lakes, *then* select temperature and observation information, and *then* add a column for temperature in Fahrenheit: 

```{r}
NTL.phys.data.processed <- 
  NTL.phys.data %>%
  filter(lakename == "Paul Lake" | lakename == "Peter Lake") %>%
  select(lakename, sampledate:temperature_C) %>%
  mutate(temperature_F = (temperature_C*9/5) + 32)
  
# If want to create a new north temperate lakes data frame, can use pipe to say "then." First take physical data set, then filter for a certain set of lakes "either peter or paul," then select different columns "lake name and sample date thru temp" and then mutate data frame to add a new column with temps in F degrees.
# Just specify which data frame you want to pull out info from "ntl.phys.data" 
# All operations happen at once when using pipe operator
# Pipe operator: %>% and says use this data to filter, select, and mutate
```

Notice that we did not place the dataset name inside the wrangling function but rather at the beginning.

### Saving processed datasets

```{r}
write.csv(NTL.phys.data.PeterPaul1, row.names = FALSE, file = "./Data/Processed/NTL-LTER_Lake_ChemistryPhysics_PeterPaul_Processed.csv")

# Row names = false, makes it so it doesn't create another column that just has the row numbers. Not a big deal, but cleans it up

# Have to load some packages in a certain order, if this is the case, close out of R and restart with next lesson that necessitates different package
```

## Closing Discussion

When we wrangle a raw dataset into a processed dataset, we create a code file that contains only the wrangling code. We then save the processed dataset as a new spreadsheet and then create a separate code file to analyze and visualize the dataset. Why do we keep the wrangling code separate from the analysis code?
