==> Lessons/13_GLMs.Rmd <==
---
title: "13: Generalized Linear Models (ANCOVA and mixed effects)"
author: "Environmental Data Analytics | Kateri Salk"
date: "Spring 2020"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
2. Apply special cases of the GLM (ANCOVA, mixed effects models) to real datasets
3. Interpret and report the results of linear regressions in publication-style formats
3. Apply model selection methods to choose model formulations

## Set up
```{r, message = FALSE}
getwd()
library(tidyverse)
library(lubridate)
library(viridis)
#install.packages("nlme")
library(nlme)
#install.packages("piecewiseSEM")
library(piecewiseSEM)

PeterPaul.chem.nutrients <- read.csv("./Data/Processed/NTL-LTER_Lake_Chemistry_Nutrients_PeterPaul_Processed.csv")
NTL.chem <- read.csv("./Data/Raw/NTL-LTER_Lake_ChemistryPhysics_Raw.csv")

NTL.chem$sampledate <- as.Date(NTL.chem$sampledate, format = "%m/%d/%y")

# Set theme
mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```

## ANCOVA
Analysis of Covariance consists of a prediction of a continuous response variable by both continuous and categorical explanatory variables. We set this up in R with the `lm` function, just like prior applications in this lesson. 

Let's say we wanted to predict total nitrogen concentrations by depth and by lake. We could represent these explanatory variables as main effects (two intercepts, same slope) or as interaction effects (two intercepts and two slopes). 

```{r, fig.height = 3, fig.width = 4}
# main effects
TNancova.main <- lm(data = PeterPaul.chem.nutrients, tn_ug ~ lakename + depth)
summary(TNancova.main)

#lakename = intercept = alpha, depth = slope = beta
# continuous = beta, only gives one beta value for each
# intercept = Paul lake at 0 meters depth with 284 micro grams/L
# if it's peter lake, the prediction for TN at 0m depth is 284 + 66
# for main effect, just giving us one estimate of slope for whatever lake it is. So, for depth, every increase in m is a 68 microgram/L increase

# interaction effects
TNancova.interaction <- lm(data = PeterPaul.chem.nutrients, tn_ug ~ lakename * depth)
summary(TNancova.interaction)

# supposing that we may have different intercepts & slopes for each lake. 
# should look the same except * instead of +
# Have 4 coeffs in model, all are predicting significant effect
# Intercept = Paul lake at depth of 0 = surface level is 228
# LakenamePeter: 228 + 161 for every 1m increase in depth 
# If the lake is paul lake, for each increase in depth of 1m getting increase of 90, but if in Peter Lake, getting decrease of -35

TNplot <- ggplot(PeterPaul.chem.nutrients, aes(x = depth, y = tn_ug, color = lakename)) + 
  geom_point(alpha = .5) + 
  geom_smooth(method = "lm", se = FALSE) +
  scale_color_manual(values = c("darkblue", "lightblue")) +
   #scale_color_brewer(palette = "Dark2") +
  labs( x = "Depth (m)", y = expression(TN~ (mu*g / L)), color = "") +
  xlim(0, 10)
print(TNplot)

# can also specify breaks in axes if don't want it to be increments of 2.5

# We have Paul with lower TN at surface that goes up as we move down in depth
# Peter, start off with higher concentration at surface and goes down at deeper depths
# Significant difference

```

## HIERARCHICAL MODELS

**Hierarchical models,** or **mixed-effects models,** are a type of linear model in which explanatory variables are given a model whose parameters are also estimated by the data. The coefficients associated with explanatory variables thus may not be a single value but instead be sampled from a distribution, called the hyper-distribution, which is defined by the modeler. The advantage of the hierarchical model is that it builds capacity to describe multiple layers of stochasticity, which enables accounting of all aspects of uncertainty in a system. Specifically, we can separately model the process of interest and the sampling process. 

Have predictory variables that can be explained with random and fixed effects
******What the coeff is on that model to help predict why: fixed
*****If know something introduces variability but don't care about specific effect: random
**** Random effect has to be categorical, can't be continuous

The coefficients of a hierarchical model are divided into two categories: **fixed effects** and **random effects.** A **fixed effect** is a factor whose levels are experimentally determined or whose interest lies in the effects of each level (e.g., covariates, treatments, interactions). A **random effect** is a factor whose levels are sampled from a larger population, or whose interest lies in the variation among them rather than the specific effect of each level. In choosing whether you are dealing with a fixed or a random effect, consider the following questions: 

  + Do you have a particular interest in the studied factor level?

  + Have you included all possible levels in the study? If want to extrapolate beyond my studied factor levels and use the factor levels to account for variability 
  
  + Do you have interest in the variance among levels?
  
  + Do you have interest in generalizing to factor levels that you did not study?
  
A common variable in hierarchical models is **space.** In many situations, we may want to infer conditions beyond the sites that we have sampled. By treating space as a random variable, we may be able to extrapolate conditions of the response variable across a spatial gradient. 

Let's think about the situation of temperature monitoring in the NTL-LTER lakes. We might be interested to know whether surface temperatures in July have increased over time in response to climate change. However, we know that there may be variability across lakes that may obscure the trend we see in temperature. We can set lake as a random effect to account for the across-lake variability and also enable us to extrapolate across lakes in northern Wisconsin.

Have summer levels changed over last 30 years? Not specific lake but lakes in general are and use that info to put into model to characterize for variability 

Let's wrangle our data and visualize a preliminary relationship between our variables of interest.

```{r}
NTL.summertemp <- 
  NTL.chem %>%
  select(lakename:temperature_C) %>% # select columns you want to look at
  #filter for Julian days in July and surface measurements
  filter(daynum > 181 & daynum < 213 & depth == 0 ) %>% #filter for just days in july & surface measurements
  #code won't work if there are NAs
  na.exclude()

NTLtemps <-
ggplot(NTL.summertemp, aes(x = sampledate, y = temperature_C, color = lakename)) +
  geom_point() +
  scale_color_viridis_d()
print(NTLtemps)

# might see little increase, but hard to see because of all the variability. Could characterize some variability to understand what influences it.
```

Next, we will build a hierarchical model. We will use the package `nlme` for our analyses. Another good package for running hierarchical, or mixed-effects, models is `lme4`. For the basic types of hierarchical models, these packages have about the same functionality. We will set year year (continuous) as a fixed effect and lake (categorical) as a random effect. Remember that we are interested in assessing if summer surface temperatures have increased in response to climate change and to account for the inter-lake variability within the model.


```{r}

# Setting year as main effect in model (how temp changes by year, and then set variability across lakes as random)

TempTest.mixed <- lme(data = NTL.summertemp,
                     temperature_C ~ year4, # Temp by year = 1 main effect
                     random = ~1|lakename) # this is our random variable (1 grouped by lakename)
summary(TempTest.mixed)
rsquared(TempTest.mixed)

# Summary from mix-effects model have dif coeffs than fixed (just temp by year). Want to add a line of best fit with Mixed instead of Fixed

# Compare the random effects model with the fixed effects model. Getting slighly different output with glm (coefficients and info with lme model)
TempTest.fixed <- gls(data = NTL.summertemp,
                      temperature_C ~ year4)
summary(TempTest.fixed)

anova(TempTest.mixed, TempTest.fixed)

# IS there a sig different fit, and if yes, which has the lower AIC? in this case we see the fixed model has the lowest 

# The lower the AIC, the better.
# The p-value tells us whether those models have a significantly different fit

NTL.tempmodel <-
ggplot(NTL.summertemp, aes(x = year4, y = temperature_C, color = lakename)) +
  geom_point() +
  scale_color_viridis() +
  geom_abline(intercept = -97.72, slope = 0.06) + # Exact coeffs of that mixed effects model
  # make it look better
  labs(x = "", y = expression("Temperature " ( degree~C)), color = "") +
  theme(legend.spacing.x = unit(0, "cm"))
print(NTL.tempmodel)

# This is how line of best fits for both data would show, but would use the one above
NTL.tempmodel <-
ggplot(NTL.summertemp, aes(x = year4, y = temperature_C)) +
  geom_point(aes(color = lakename)) +
  scale_color_viridis_d() +
  geom_smooth(method = "lm", se = FALSE) +
  geom_abline(intercept = -97.72, slope = 0.06) + # Exact coeffs of that mixed effects model
  # make it look better
  labs(x = "", y = expression("Temperature " ( degree~C)), color = "") +
  theme(legend.spacing.x = unit(0, "cm"))
print(NTL.tempmodel)

# Black = mixed effects model, Blue = Fixed effects model 

# Slope shows: there was a sig change in temp, because slope was sig. A sig predictor in model may not be environmentally relevent but means for each year, seeing increase of .06 degrees C. If say .06 * 30 = surface water degrees in Wisconsin lakes have increased by 1.8 degrees C in last 30 years

# Theme edits made all the labels for lakes closer together
```

Question: How would you interpret the collective results of your mixed effects model in the context of the study question?

> ANSWER:
